Drizzle Version 1
We've done it, the website is up. A sample set of notes, rainfalls, and rain jars can be navigated through and interacted with. Javascript menu switching allows for a compact one page site. Sample forms for assign to the database have been created but not implemented. Next step is making the site able to take in and display inputs.